backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
sentence-transformers==3.0.0
pinecone-client==3.2.2
spacy==3.7.4
numpy==1.26.4
scikit-learn==1.4.2
python-dotenv==1.0.1
python-multipart==0.0.9
httpx==0.27.0
groq==0.8.0
pytest==8.2.0
pytest-asyncio==0.23.7
